requirements-dev.txt
-r requirements.txt
jupyter
django-extensions==3.1.3
django-test-without-migrations==0.6